move comoonents to src
