+++ client/app/+not-found.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link, Stack } from 'expo-router';
 import { StyleSheet } from 'react-native';
 
-import { Text, View } from '@/components/Themed';
+import { Text, View } from '@/src/components/Themed';
 
 export default function NotFoundScreen() {
   return (
