+++ client/src/components/ErrorScreen.tsx
