+++ client/src/components/StyledText.tsx
